+++ 3ds/build.gradle
@@ -1,0 +1,35 @@
+plugins {
+    id 'com.android.library' apply false
+    id 'kotlin-android' apply false
+}
+
+apply from: "$rootProject.projectDir/android-module.gradle"
+
+def DEVICE_FINGERPRINT_ORG_ID = '"1snn5n9w"'
+def DEVICE_FINGERPRINT_MERCHANT_ID = '"airwallex_cybs"'
+
+android {
+    namespace 'com.airwallex.android.threedsecurity'
+}
+
+android.buildTypes.each { type ->
+    type.buildConfigField 'String', 'DEVICE_FINGERPRINT_ORG_ID', DEVICE_FINGERPRINT_ORG_ID
+    type.buildConfigField 'String', 'DEVICE_FINGERPRINT_MERCHANT_ID', DEVICE_FINGERPRINT_MERCHANT_ID
+}
+
+dependencies {
+    api project(':ui-core')
+
+    // fingerprinting
+    implementation 'io.github.airwallex:TrustDefender-Behavioral:6.1-67'
+    implementation 'io.github.airwallex:TrustDefender-Profiling:6.1-67'
+    implementation 'io.github.airwallex:TrustDefender-Profiling-Connections:6.1-67'
+
+    // test
+    testImplementation "junit:junit:$junitVersion"
+    testImplementation "org.robolectric:robolectric:$robolectricVersion"
+    testImplementation "androidx.test:core:$testCoreVersion"
+    testImplementation "org.jetbrains.kotlin:kotlin-test-junit:$kotlinVersion"
+}
+
+apply from: rootProject.file("publish-module.gradle")
