+++ 3ds/gradle.properties
@@ -1,0 +1,3 @@
+MAVEN_ARTIFACT_ID=payment-3ds
+MAVEN_ARTIFACT_NAME=Airwallex Payment Three Dimension Security
+MAVEN_ARTIFACT_DESCRIPTION=Airwallex Payment Three Dimension Security SDK
