+++ 3ds/src/main/java/com/airwallex/android/threedsecurity/exception/ThreeDSException.kt
@@ -1,0 +1,7 @@
+package com.airwallex.android.threedsecurity.exception
+
+import com.airwallex.android.core.exception.AirwallexException
+
+class ThreeDSException(
+    message: String
+) : AirwallexException(null, null, 0, message)
