+++ 3ds/src/main/java/com/airwallex/android/threedsecurity/exception/WebViewConnectionException.kt
@@ -1,0 +1,9 @@
+package com.airwallex.android.threedsecurity.exception
+
+import com.airwallex.android.core.exception.AirwallexException
+import com.airwallex.android.core.model.AirwallexError
+
+class WebViewConnectionException @JvmOverloads constructor(
+    message: String,
+    e: Throwable? = null
+) : AirwallexException(AirwallexError(message = message), null, 0, message, e)
