+++ publish-module.gradle
@@ -94,6 +94,9 @@ project.afterEvaluate {
                                     case 'components-core':
                                         dependencyNode.appendNode('artifactId', 'payment-components-core')
                                         break
+                                    case '3ds':
+                                        dependencyNode.appendNode('artifactId', 'payment-3ds')
+                                        break
                                     default:
                                         break
                                 }
