+++ sample/build.gradle
@@ -106,6 +106,7 @@ dependencies {
 //    implementation 'io.github.airwallex:payment-card:4.1.0'
 //    implementation 'io.github.airwallex:payment-redirect:4.1.0'
 //    implementation 'io.github.airwallex:payment-wechat:4.1.0'
+//    implementation 'io.github.airwallex:payment-googlepay:4.1.0'
 
     implementation 'com.google.android.material:material:1.4.0'
     implementation 'androidx.preference:preference-ktx:1.1.1'
