+++ sample/src/main/res/values/arrays.xml
@@ -22,6 +22,11 @@
         <item>True</item>
     </string-array>
 
+    <string-array name="array_force_3ds">
+        <item>False</item>
+        <item>True</item>
+    </string-array>
+
     <string-array name="array_auto_capture">
         <item>Enabled</item>
         <item>Disabled</item>
