+++ sample/src/main/res/values/strings.xml
@@ -63,6 +63,7 @@
     <string name="generate_customer_failed">Generate customer failed</string>
     <string name="generate_customer_failed_message">There was an error while generate customer. Please try again.</string>
     <string name="requires_cvc">Requires CVC</string>
+    <string name="force_3ds">Force 3DS</string>
     <string name="auto_capture">Auto Capture</string>
     <string name="shipping">Shipping</string>
     <string name="item">Item</string>
